+++ crates/types/src/api/v1/rooms/mod.rs
@@ -4,9 +4,9 @@
 
 //! This module contains types that are used for OpenTalk API V1 rooms endpoints.
 
+mod delete_room_query;
 mod get_room_event_request;
 mod get_room_event_response;
-mod internal_room_delete_query;
 mod patch_rooms_request_body;
 mod post_rooms_request_body;
 mod post_rooms_start_invited_request;
@@ -20,9 +20,9 @@ mod start_room_error;
 pub mod sip_config_resource;
 pub mod streaming_targets;
 
+pub use delete_room_query::DeleteRoomQuery;
 pub use get_room_event_request::GetRoomEventRequest;
 pub use get_room_event_response::GetRoomEventResponse;
-pub use internal_room_delete_query::DeleteRoomQuery;
 pub use patch_rooms_request_body::PatchRoomsRequestBody;
 pub use post_rooms_request_body::PostRoomsRequestBody;
 pub use post_rooms_start_invited_request::PostRoomsStartInvitedRequest;